Function fixes and code cleanup
Added some comments to explain a bit of each function and what they do

Fixed functions sheetDF and searchDate() to match be able to answer to every possible situation (different start and end months or same start and end months)

created sheetInfo() and getDateFormat(), because it was identified that those portions of the script were repeated in some instances.

Created writeTXT() The function that should generate a TXT file with the schedule info. THIS FUNCTION IS NOT FINISHED